% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_obis}
\alias{fetch_obis}
\title{Fetch the complete OBIS dataset}
\usage{
fetch_obis(
  uri = query_obis(),
  path = get_root(),
  timeout = 600,
  unpack = TRUE,
  ...
)
}
\arguments{
\item{uri}{charcater, the URL of the file to fetch - see \code{query_obis}}

\item{path}{character, the path (directory to save data to)}

\item{timeout}{integer, the timeout to set (system default is 60s, but
we set default here to 600s)}

\item{unpack}{logical, if TRUE unzip the file which will overwrite prior
instances of 'occurrence.csv'}

\item{...}{other arguments for \code{download.file}}
}
\value{
named integer, 0 for success, name is of the output file
}
\description{
Fetch the complete OBIS dataset
}
