% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullobis-package.R
\docType{package}
\name{fullobis-package}
\alias{fullobis-package}
\alias{fullobis}
\title{fullobis}
\description{
Provides simplified access to \href{https://obis.org/manual/access}{OBIS} data in parquet format.
}
