% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_current}
\alias{get_current}
\title{Retrieve the most recent parquet filename in the root path}
\usage{
get_current(root = get_root(), pattern = "^obis_.*\\\\.parquet$")
}
\arguments{
\item{root}{character, the root path to the dataset}

\item{pattern}{the filename pattern to search, by default we use the equivalent of the glob \code{obis_*.parquet}}
}
\value{
the fully qualified filename and path or "" if none found
}
\description{
Retrieve the most recent parquet filename in the root path
}
