% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_root}
\alias{get_root}
\title{Retrieve the root directory from options}
\usage{
get_root(default = ".")
}
\arguments{
\item{default}{character, if the option is not set then return this as the default}
}
\value{
charcater path to the root directory
}
\description{
Retrieve the root directory from options
}
