% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_useragent}
\alias{get_useragent}
\title{Retrieve the useragent from options}
\usage{
get_useragent(default = "")
}
\arguments{
\item{default}{character, if the option is not set then return this as the default}
}
\value{
charcater user agent for \code{\link[httr]{user_agent}}
}
\description{
Retrieve the useragent from options
}
