% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_root}
\alias{set_root}
\title{Set the root directory in session options}
\usage{
set_root(path = ".")
}
\arguments{
\item{path}{character, the root path to setthe root directory}
}
\value{
see \code{options}
}
\description{
Set the root directory in session options
}
