% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_useragent}
\alias{set_useragent}
\title{Set the useragent in session options}
\usage{
set_useragent(useragent = "")
}
\arguments{
\item{useragent}{character, the useragent to set}
}
\value{
see \code{options}
}
\description{
Set the useragent in session options
}
